+++ examples/todos/main.py
@@ -1,59 +1,25 @@
 import logging
-import os
 from contextlib import asynccontextmanager
 from typing import Annotated
 
+from beanie import PydanticObjectId
 from fastapi import Depends, FastAPI, HTTPException, Query, status
-from models import (
-    MongoTodo,
-    MongoTodoList,
-    RedisTodo,
-    RedisTodoList,
-    SqlTodo,
-    SqlTodoList,
-)
+from models import MongoTodoList, RedisTodoList, SqlTodoList
 from schemas import TodoList
+from stores import clear_stores, get_mongo_store, get_redis_store, get_sql_store
 
-from nqlstore import (
-    MongoStore,
-    RedisStore,
-    SQLStore,
-)
+from nqlstore import MongoStore, RedisStore, SQLStore
 
-_SQL_STORE: SQLStore | None = None
-_REDIS_STORE: RedisStore | None = None
-_MONGO_STORE: MongoStore | None = None
-
-# dependencies
-_SqlStoreDep = Annotated[SQLStore | None, Depends(lambda: _SQL_STORE)]
-_RedisStoreDep = Annotated[RedisStore | None, Depends(lambda: _REDIS_STORE)]
-_MongoStoreDep = Annotated[MongoStore | None, Depends(lambda: _MONGO_STORE)]
+_SqlStoreDep = Annotated[SQLStore | None, Depends(get_sql_store)]
+_RedisStoreDep = Annotated[RedisStore | None, Depends(get_redis_store)]
+_MongoStoreDep = Annotated[MongoStore | None, Depends(get_mongo_store)]
 
 
-# startup events
 @asynccontextmanager
 async def lifespan(app_: FastAPI):
-    sql_url = os.environ.get("SQL_URL", "")
-    redis_url = os.environ.get("REDIS_URL", "")
-    mongo_url = os.environ.get("MONGO_URL", "")
-    mongo_db = os.environ.get("MONGO_DB", "todos")
-
-    if sql_url:
-        global _SQL_STORE
-        _SQL_STORE = SQLStore(uri=sql_url)
-        await _SQL_STORE.register([SqlTodoList, SqlTodo])
-
-    if redis_url:
-        global _REDIS_STORE
-        _REDIS_STORE = RedisStore(uri=redis_url)
-        await _REDIS_STORE.register([RedisTodoList, RedisTodo])
-
-    if mongo_url:
-        global _MONGO_STORE
-        _MONGO_STORE = MongoStore(uri=mongo_url, database=mongo_db)
-        await _MONGO_STORE.register([MongoTodoList, MongoTodo])
-
+    clear_stores()
     yield
+    clear_stores()
 
 
 app = FastAPI(lifespan=lifespan)
@@ -67,15 +33,17 @@ async def search(
     q: str = Query(...),
 ):
     """Searches for todos by name"""
+
     results = []
-    query = {"name": {"$in": q}}
+    query = {"name": {"$regex": f".*{q}.*", "$options": "i"}}
 
     try:
         if sql:
             results += await sql.find(SqlTodoList, query=query)
 
         if redis:
-            results += await redis.find(RedisTodoList, query=query)
+            # redis's regex search is not mature so we use its full text search
+            results += await redis.find(RedisTodoList, (RedisTodoList.name % f"*{q}*"))
 
         if mongo:
             results += await mongo.find(MongoTodoList, query=query)
@@ -83,19 +51,19 @@ async def search(
         logging.error(exp)
         raise exp
 
-    return results
+    return [item.model_dump(mode="json") for item in results]
 
 
-@app.get("/todos/{id}")
+@app.get("/todos/{id_}")
 async def get_one(
     sql: _SqlStoreDep,
     redis: _RedisStoreDep,
     mongo: _MongoStoreDep,
-    id: int | str,
+    id_: int | str,
 ):
     """Get todolist by id"""
     results = []
-    query = {"id": {"$eq": id}}
+    query = {"id": {"$eq": id_}}
 
     try:
         if sql:
@@ -105,13 +73,15 @@ async def get_one(
             results += await redis.find(RedisTodoList, query=query, limit=1)
 
         if mongo:
-            results += await mongo.find(MongoTodoList, query=query, limit=1)
+            results += await mongo.find(
+                MongoTodoList, query={"_id": {"$eq": PydanticObjectId(id_)}}, limit=1
+            )
     except Exception as exp:
         logging.error(exp)
         raise exp
 
     try:
-        return results[0]
+        return results[0].model_dump(mode="json")
     except IndexError:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
 
@@ -137,23 +107,24 @@ async def create_one(
         if mongo:
             results += await mongo.insert(MongoTodoList, [payload_dict])
 
-        return results[0]
+        result = results[0].model_dump(mode="json")
+        return result
     except Exception as exp:
         logging.error(exp)
         raise exp
 
 
-@app.put("/todos/{id}")
+@app.put("/todos/{id_}")
 async def update_one(
     sql: _SqlStoreDep,
     redis: _RedisStoreDep,
     mongo: _MongoStoreDep,
-    id: int | str,
+    id_: int | str,
     payload: TodoList,
 ):
     """Update a todolist"""
     results = []
-    query = {"id": {"$eq": id}}
+    query = {"id": {"$eq": id_}}
     updates = payload.model_dump(exclude_unset=True)
 
     try:
@@ -164,42 +135,48 @@ async def update_one(
             results += await redis.update(RedisTodoList, query=query, updates=updates)
 
         if mongo:
-            results += await mongo.update(MongoTodoList, query=query, updates=updates)
+            results += await mongo.update(
+                MongoTodoList,
+                query={"_id": {"$eq": PydanticObjectId(id_)}},
+                updates=updates,
+            )
     except Exception as exp:
         logging.error(exp)
         raise exp
 
     try:
-        return results[0]
+        return results[0].model_dump(mode="json")
     except IndexError:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
 
 
-@app.delete("/todos/{id}")
+@app.delete("/todos/{id_}")
 async def delete_one(
     sql: _SqlStoreDep,
     redis: _RedisStoreDep,
     mongo: _MongoStoreDep,
-    id: int | str,
+    id_: int | str,
 ):
     """Delete a todolist"""
     results = []
-    query = {"id": {"$eq": id}}
+    query = {"id": {"$eq": id_}}
 
     try:
         if sql:
             results += await sql.delete(SqlTodoList, query=query)
 
         if redis:
-            results += await redis.find(RedisTodoList, query=query)
+            results += await redis.delete(RedisTodoList, query=query)
 
         if mongo:
-            results += await mongo.find(MongoTodoList, query=query)
+            results += await mongo.delete(
+                MongoTodoList, query={"_id": {"$eq": PydanticObjectId(id_)}}
+            )
     except Exception as exp:
         logging.error(exp)
         raise exp
 
     try:
-        return results[0]
+        return results[0].model_dump(mode="json")
     except IndexError:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)